Refactored config retriever
