+++ tsdat/config/retriever/data_converter_config.py
@@ -1,0 +1,7 @@
+from pydantic import Extra
+
+from ..utils import ParameterizedConfigClass
+
+
+class DataConverterConfig(ParameterizedConfigClass, extra=Extra.allow):
+    ...
