+++ tsdat/config/retriever/data_reader_config.py
@@ -1,0 +1,5 @@
+from ..utils import ParameterizedConfigClass
+
+
+class DataReaderConfig(ParameterizedConfigClass):
+    ...
